Network.Lastfm.Album

Description

Lastfm album API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Album as Album

Documentation

class ArtistAlbumOrMBID a

Unify (Artist → Album → …) and (MBID → …)

addTags :: Request f Sign (Artist -> Album -> [Tag] -> APIKey -> SessionKey -> Ready)

Tag an album using a list of user supplied tags.

http://www.last.fm/api/show/album.addTags

getBuyLinks :: ArtistAlbumOrMBID t => Request f Send (Country -> t)

Get a list of Buy Links for a particular Album. It is required that you supply either the artist and track params or the mbid parameter.

Optional: autocorrect

http://www.last.fm/api/show/album.getBuylinks

getInfo :: ArtistAlbumOrMBID t => Request f Send t

Get the metadata for an album on Last.fm using the album name or a musicbrainz id. See playlist.fetch on how to get the album playlist.

Optional: autocorrect, username, language

http://www.last.fm/api/show/album.getInfo

getShouts :: ArtistAlbumOrMBID t => Request f Send t

Get shouts for this album.

Optional: autocorrect, limit, page

http://www.last.fm/api/show/album.getShouts

getTags :: ArtistAlbumOrMBID t => Request f a t

Get the tags applied by an individual user to an album on Last.fm.

Optional: autocorrect, user

http://www.last.fm/api/show/album.getTags

getTopTags :: ArtistAlbumOrMBID t => Request f Send t

Get the top tags for an album on Last.fm, ordered by popularity.

Optional: autocorrect

http://www.last.fm/api/show/album.getTopTags

removeTag :: Request f Sign (Artist -> Album -> Tag -> APIKey -> SessionKey -> Ready)

Remove a user's tag from an album.

http://www.last.fm/api/show/album.removeTag

search :: Request f Send (Album -> APIKey -> Ready)

Search for an album by name. Returns album matches sorted by relevance.

Optional: limit, page

http://www.last.fm/api/show/album.search

share :: Request f Sign (Album -> Artist -> Recipient -> APIKey -> SessionKey -> Ready)

Share an album with one or more Last.fm users or other friends.

Optional: public, message, recipient

http://www.last.fm/api/show/album.share