Network.Lastfm.Artist

Description

Lastfm artist API

This module is intended to be imported qualified:

import qualified Network.Lastfm.Artist as Artist

Documentation

class ArtistOrMBID r a

Unify (Artist -> …) and (MBID -> …)

addTags :: Request f (Artist -> [Tag] -> APIKey -> SessionKey -> Sign)

Tag an artist with one or more user supplied tags.

http://www.last.fm/api/show/artist.addTags

getCorrection :: Request f (Artist -> APIKey -> Ready)

Use the last.fm corrections data to check whether the supplied artist has a correction to a canonical artist

http://www.last.fm/api/show/artist.getCorrection

getEvents :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get a list of upcoming events for this artist. Easily integratable into calendars, using the ical standard (see feeds section below).

Optional: autocorrect, limit, pages, festivalsonly

http://www.last.fm/api/show/artist.getEvents

getInfo :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get the metadata for an artist. Includes biography.

Optional: language, autocorrect, username

http://www.last.fm/api/show/artist.getInfo

getPastEvents :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get a paginated list of all the events this artist has played at in the past.

Optional: page, autocorrect, limit

http://www.last.fm/api/show/artist.getPastEvents

getPodcast :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get a podcast of free mp3s based on an artist

Optional: autocorrect

http://www.last.fm/api/show/artist.getPodcast

getShouts :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get shouts for this artist. Also available as an rss feed.

Optional:autocorrect, limit, page

http://www.last.fm/api/show/artist.getShouts

getSimilar :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get all the artists similar to this artist

Optional: limit, autocorrect

http://www.last.fm/api/show/artist.getSimilar

getTags :: ArtistOrMBID r a => Request f (a -> APIKey -> r)

Get the tags applied by an individual user to an artist on Last.fm. If accessed as an authenticated service and you don't supply a user parameter then this service will return tags for the authenticated user. To retrieve the list of top tags applied to an artist by all users use getTopTags.

Optional: user, autocorrect

http://www.last.fm/api/show/artist.getTags

getTopAlbums :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get the top albums for an artist on Last.fm, ordered by popularity.

Optional: autocorrect, page, limit

http://www.last.fm/api/show/artist.getTopAlbums

getTopFans :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get the top fans for an artist on Last.fm, based on listening data.

Optional: autocorrect

http://www.last.fm/api/show/artist.getTopFans

getTopTags :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get the top tags for an artist on Last.fm, ordered by popularity.

Optional: autocorrect

http://www.last.fm/api/show/artist.getTopTags

getTopTracks :: ArtistOrMBID Ready a => Request f (a -> APIKey -> Ready)

Get the top tracks by an artist on Last.fm, ordered by popularity

Optional: autocorrect, page, limit

http://www.last.fm/api/show/artist.getTopTracks

removeTag :: Request f (Artist -> Tag -> APIKey -> SessionKey -> Sign)

Remove a user's tag from an artist.

http://www.last.fm/api/show/artist.removeTag

search :: Request f (Artist -> APIKey -> Ready)

Search for an artist by name. Returns artist matches sorted by relevance.

Optional: limit, page

http://www.last.fm/api/show/artist.search

share :: Request f (Artist -> Recipient -> APIKey -> SessionKey -> Sign)

Share an artist with Last.fm users or other friends.

Optional: message, public

http://www.last.fm/api/show/artist.share

shout :: Request f (Artist -> Message -> APIKey -> SessionKey -> Sign)

Shout in this artist's shoutbox

http://www.last.fm/api/show/artist.shout