Network.Lastfm.Geo

Description

Lastfm geo API

This module is intended to be imported qualified:

import qualified Network.Lastfm.Geo as Geo

Documentation

getEvents :: Request f (APIKey -> Ready)

Get all events in a specific location by country or city name.

Optional: longitude, latitude, location, distance, page, tag, festivalsonly, limit

http://www.last.fm/api/show/geo.getEvents

getMetroArtistChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of artists for a metro

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroArtistChart

getMetroHypeArtistChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of hyped (up and coming) artists for a metro

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroHypeArtistChart

getMetroHypeTrackChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of tracks for a metro

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroHypeTrackChart

getMetroTrackChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of tracks for a metro

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroTrackChart

getMetroUniqueArtistChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of the artists which make that metro unique

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroUniqueArtistChart

getMetroUniqueTrackChart :: Request f (Metro -> Country -> APIKey -> Ready)

Get a chart of tracks for a metro

Optional: start, end, page, limit

http://www.last.fm/api/show/geo.getMetroUniqueTrackChart

getMetroWeeklyChartlist :: Request f (Metro -> APIKey -> Ready)

Get a list of available chart periods for this metro, expressed as date ranges which can be sent to the chart services.

http://www.last.fm/api/show/geo.getMetroWeeklyChartlist

getMetros :: Request f (APIKey -> Ready)

Get a list of valid countries and metros for use in the other webservices

Optional: country

http://www.last.fm/api/show/geo.getMetros

getTopArtists :: Request f (Country -> APIKey -> Ready)

Get the most popular artists on Last.fm by country

Optional: limit, page

http://www.last.fm/api/show/geo.getTopArtists

getTopTracks :: Request f (Country -> APIKey -> Ready)

Get the most popular tracks on Last.fm last week by country

Optional: limit, page

http://www.last.fm/api/show/geo.getTopTracks