Network.Lastfm.Group

Description

Lastfm group API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Group as Group

Documentation

getHype :: Request f Send (Group -> APIKey -> Ready)

Get the hype list for a group

http://www.last.fm/api/show/group.getHype

getMembers :: Request f Send (Group -> APIKey -> Ready)

Get a list of members for this group.

Optional: page, limit

http://www.last.fm/api/show/group.getMembers

getWeeklyAlbumChart :: Request f Send (Group -> APIKey -> Ready)

Get an album chart for a group, for a given date range. If no date range is supplied, it will return the most recent album chart for this group.

Optional: from, to

http://www.last.fm/api/show/group.getWeeklyAlbumChart

getWeeklyArtistChart :: Request f Send (Group -> APIKey -> Ready)

Get an artist chart for a group, for a given date range. If no date range is supplied, it will return the most recent album chart for this group.

Optional: from, to

http://www.last.fm/api/show/group.getWeeklyArtistChart

getWeeklyChartList :: Request f Send (Group -> APIKey -> Ready)

Get a list of available charts for this group, expressed as date ranges which can be sent to the chart services.

http://www.last.fm/api/show/group.getWeeklyChartList

getWeeklyTrackChart :: Request f Send (Group -> APIKey -> Ready)

Get a track chart for a group, for a given date range. If no date range is supplied, it will return the most recent album chart for this group.

Optional: from, to

http://www.last.fm/api/show/group.getWeeklyTrackChart