Network.Lastfm.Internal

Description

liblastfm internals

You shouldn't need to import this module unless you are doing something interesting.

Documentation

newtype Request f a

Lastfm API request data type

a is the authentication state. Can be Ready, which means this Request is ready to be sent, or Sign, if the request signature hasn't been computed yet

f is the response format (liblastfm supports both JSON and XML)

Constructors

Request

Fields

unRequest :: Const (Dual (Endo (R f))) a

data Format

Response format: either JSON or XML

Constructors

JSON

XML

data Ready

Request that is ready to be sent

data Sign

Request that requires signing procedure

data R f

Lastfm API request data type

low-level representation

Constructors

R

Fields

_host :: !Text

_method :: !ByteString

_query :: !(Map Text Text)

wrap :: (R f -> R f) -> Request f a

Wrapping to interesting Monoid (R -> R) instance

unwrap :: Request f a -> R f -> R f

Unwrapping from interesting Monoid (R -> R) instance

render :: R f -> String

Construct String from request for networking

coerce :: Request f a -> Request f b

absorbQuery :: Foldable t => t (Request f b) -> Request f a

Absorbing a bunch of queries, useful in batch operations

indexedWith :: Int -> Request f a -> Request f a

Transforming Request to the "array notation"

Lenses

host :: Functor f => (Text -> f Text) -> R h -> f (R h)

Request _host

method :: Functor f => (ByteString -> f ByteString) -> R h -> f (R h)

Request HTTP _method

query :: Functor f => (Map Text Text -> f (Map Text Text)) -> R h -> f (R h)

Request _query string