Network.Lastfm.Library

Description

Lastfm library API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Library as Library

Documentation

addAlbum :: Request f Sign (Artist -> Album -> APIKey -> SessionKey -> Ready)

Add an album or collection of albums to a user's Last.fm library

http://www.last.fm/api/show/library.addAlbum

addArtist :: Request f Sign (Artist -> APIKey -> SessionKey -> Ready)

Add an artist to a user's Last.fm library

http://www.last.fm/api/show/library.addArtist

addTrack :: Request f Sign (Artist -> Track -> APIKey -> SessionKey -> Ready)

Add a track to a user's Last.fm library

http://www.last.fm/api/show/library.addTrack

getAlbums :: Request f Send (User -> APIKey -> Ready)

A paginated list of all the albums in a user's library, with play counts and tag counts.

Optional: artist, limit, page

http://www.last.fm/api/show/library.getAlbums

getArtists :: Request f Send (User -> APIKey -> Ready)

A paginated list of all the artists in a user's library, with play counts and tag counts.

Optional: limit, page

http://www.last.fm/api/show/library.getArtists

getTracks :: Request f Send (User -> APIKey -> Ready)

A paginated list of all the tracks in a user's library, with play counts and tag counts.

Optional: artist, album, page, limit

http://www.last.fm/api/show/library.getTracks

removeAlbum :: Request f Sign (Artist -> Album -> APIKey -> SessionKey -> Ready)

Remove an album from a user's Last.fm library

http://www.last.fm/api/show/library.removeAlbum

removeArtist :: Request f Sign (Artist -> APIKey -> SessionKey -> Ready)

Remove an artist from a user's Last.fm library

http://www.last.fm/api/show/library.removeArtist

removeScrobble :: Request f Sign (Artist -> Track -> Timestamp -> APIKey -> SessionKey -> Ready)

Remove a scrobble from a user's Last.fm library

http://www.last.fm/api/show/library.removeScrobble

removeTrack :: Request f Sign (Artist -> Track -> APIKey -> SessionKey -> Ready)

Remove a track from a user's Last.fm library

http://www.last.fm/api/show/library.removeTrack