Network.Lastfm.Radio

Description

Lastfm radio API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Radio as Radio

Documentation

getPlaylist :: Request f (Multiplier -> Bitrate -> APIKey -> SessionKey -> Sign)

Fetch new radio content periodically in an XSPF format.

Optional: discovery, rtp, buyLinks

http://www.last.fm/api/show/radio.getPlaylist

search :: Request f (Name -> APIKey -> Ready)

Resolve the name of a resource into a station depending on which resource it is most likely to represent.

http://www.last.fm/api/show/radio.search

tune :: Request f (Station -> APIKey -> SessionKey -> Sign)

Tune in to a Last.fm radio station.

Optional: language

http://www.last.fm/api/show/radio.tune