Network.Lastfm.Request

Description

Request construction

Request

data Request f a

Lastfm API request data type

a is authentication state. Might be Send which indicates that you may send this request already or Sign, when request signature isn't computed yet

f is response format. liblastfm currently supports JSON or XML

data R f

Lastfm API request data type

low-level representation

data Ready

Request that is ready to be sent

data Sign

Request that requires signing procedure

data Format

Response format: either JSON or XML

Constructors

JSON

XML

Request major parameters

api :: Text -> Request f a

Change request API method

Primarily used in API call wrappers, not intended for usage by library user

post :: Request f a

Change html _method to POST

Primarily used in API call wrappers, not intended for usage by library user

get :: Request f a

Change html _method to GET

Primarily used in API call wrappers, not intended for usage by library user

json :: Request JSON a

Change API response format to JSON

This is a little helper. It's actually enough to specialize Format manually

xml :: Request XML a

Change API response format to XML

This is a little helper. It's actually enough to specialize Format manually

data APIKey

apiKey :: Text -> Request f APIKey

Change request API key

data SessionKey

sessionKey :: Text -> Request f SessionKey

Change request session key

Request minor parameters

data Token

token :: Text -> Request f Token

Add token parameter

data Callback

callback :: Text -> Request f Callback

Add callback link parameter

data Artist

artist :: Text -> Request f Artist

Add artist parameter

artists :: [Text] -> Request f [Artist]

Add artist parameter

data Album

album :: Text -> Request f Album

Add artist parameter

data MBID

mbid :: Text -> Request f MBID

Add MBID parameter

data Country

country :: Text -> Request f Country

Add country parameter

data Autocorrect

autocorrect :: Bool -> Request f Autocorrect

Add autocorrect parameter

data Event

event :: Int64 -> Request f Event

Add event parameter

data Status

Constructors

Attending

Maybe

NotAttending

status :: Status -> Request f Status

Add status parameter

data From

from :: Int64 -> Request f From

Add from parameter

data To

to :: Int64 -> Request f To

Add to parameter

data Group

group :: Text -> Request f Group

Add group parameter

data Language

language :: Text -> Request f Language

Add language parameter

data Distance

distance :: Int64 -> Request f Distance

Add distance parameter

data Longitude

longitude :: Text -> Request f Longitude

Add longitude parameter

data Latitude

latitude :: Text -> Request f Latitude

Add latitude parameter

data Location

location :: Text -> Request f Location

Add location parameter

data Start

start :: Int64 -> Request f Start

Add start parameter

data End

end :: Int64 -> Request f End

Add end parameter

data Festivals

festivalsonly :: Bool -> Request f Festivals

Add festivalsonly parameter

data StartTimestamp

startTimestamp :: Int64 -> Request f StartTimestamp

Add startTimestamp parameter

data EndTimestamp

endTimestamp :: Int64 -> Request f EndTimestamp

Add endTimestamp parameter

data Metro

metro :: Text -> Request f Metro

Add metro parameter

data Tag

tags :: [Text] -> Request f [Tag]

Add tags parameter

tag :: Text -> Request f Tag

Add tag parameter

data Track

track :: Text -> Request f Track

Add track parameter

data Timestamp

timestamp :: Int64 -> Request f Timestamp

Add timestamp parameter

data Fingerprint

fingerprint :: Int64 -> Request f Fingerprint

Add fingerprint parameter

data AlbumArtist

albumArtist :: Text -> Request f AlbumArtist

Add albumArtist parameter

data Duration

duration :: Int64 -> Request f Duration

Add duration parameter

data TrackNumber

trackNumber :: Int64 -> Request f TrackNumber

Add trackNumber parameter

data Playlist

playlist :: Int64 -> Request f Playlist

Add playlistID parameter

data Title

title :: Text -> Request f Title

Add title parameter

data Description

description :: Text -> Request f Description

Add description parameter

data ChosenByUser

chosenByUser :: Bool -> Request f ChosenByUser

Add chosenByUser parameter

data Context

context :: Text -> Request f Context

Add context parameter

data StreamId

streamId :: Int64 -> Request f StreamId

Add streamId parameter

data RecentTracks

recentTracks :: Bool -> Request f RecentTracks

Add recentTracks parameter

data Recipient

recipient :: Text -> Request f Recipient

Add recipient parameter

data Username

username :: Text -> Request f Username

Add username parameter

data User

user :: Text -> Request f User

Add user parameter

data Password

password :: Text -> Request f Password

Add password parameter

data Public

public :: Bool -> Request f Public

Add public parameter

data Message

message :: Text -> Request f Message

Add message parameter

data Page

page :: Int64 -> Request f Page

Add page parameter

data Limit

limit :: Int64 -> Request f Limit

Add limit parameter

data TaggingType

taggingType :: Text -> Request f TaggingType

Add taggingType parameter

data UseRecs

useRecs :: Bool -> Request f UseRecs

Add useRecs parameter

data Venue

venue :: Int64 -> Request f Venue

Add venue parameter

data VenueName

venueName :: Text -> Request f VenueName

Add venue parameter

data Discovery

discovery :: Bool -> Request f Discovery

Add group parameter

data RTP

rtp :: Bool -> Request f RTP

Add rtp parameter

data BuyLinks

buyLinks :: Bool -> Request f BuyLinks

Add buyLinks parameter

data Multiplier

Constructors

M1

M2

multiplier :: Multiplier -> Request f Multiplier

Add multiplier parameter

data Bitrate

Constructors

B64

B128

bitrate :: Bitrate -> Request f Bitrate

Add bitrate parameter

data Name

name :: Text -> Request f Name

Add name parameter

data Station

station :: Text -> Request f Station

Add station parameter

class Targeted a

comparison :: Targeted a => Int64 -> Request f a -> Request f a

Add comparison parameter