Network.Lastfm.Response

Description

Request sending and Response parsing

Request signature

Signature is required for every authenticated API request. Basically, every such request appends the md5 footprint of its arguments to the query as described at http://www.last.fm/api/authspec#8

newtype Secret

Application secret

Constructors

Secret Text

sign :: Secret -> Request f Sign -> Request f Ready

Sign Request with Secret

Get Response

class Supported f

Supported provides parsing for a chosen Format

JSON is parsed to aeson's Value, XML is to lazy ByteString (in other words, parsing XML is left to the user)

Associated Types

type Response f

data Format

Response format: either JSON or XML

Constructors

JSON

XML

lastfm :: Supported f => Request f Ready -> IO (Either LastfmError (Response f))

Send Request and parse the Response

Also catches HttpExceptions thrown by http-conduit

lastfm_ :: Supported f => Request f Ready -> IO ()

Send Request without parsing the Response

Errors

data LastfmError

Different ways last.fm response can be unusable

Constructors

LastfmBadResponse ByteString
last.fm thinks it responded with something legible, but it really isn't

LastfmEncodedError Int
last.fm encoded error into the response and sent it along with HTTP 200 response code

LastfmHttpException HttpException
http-conduit exception, wrapped

_LastfmBadResponse :: (Choice p, Applicative m) => p ByteString (m ByteString) -> p LastfmError (m LastfmError)

This is a Prism' LastfmError ByteString in disguise

_LastfmEncodedError :: (Choice p, Applicative m) => p Int (m Int) -> p LastfmError (m LastfmError)

This is a Prism' LastfmError Int in disguise

_LastfmHttpException :: (Choice p, Applicative m) => p HttpException (m HttpException) -> p LastfmError (m LastfmError)

This is a Prism' LastfmError HttpException in disguise

Internal

lastfm' :: Supported f => (Proxy f -> ByteString -> a) -> R f -> IO a

Send R and parse Response with the supplied function

finalize :: Supported f => Request f Ready -> R f

Get R from Request