Network.Lastfm.Tag

Description

Lastfm tag API

This module is intended to be imported qualified:

import qualified Network.Lastfm.Tag as Tag

Documentation

getInfo :: Request f (Tag -> APIKey -> Ready)

Get the metadata for a tag

Optional: language

http://www.last.fm/api/show/tag.getInfo

getSimilar :: Request f (Tag -> APIKey -> Ready)

Search for tags similar to this one. Returns tags ranked by similarity, based on listening data.

http://www.last.fm/api/show/tag.getSimilar

getTopAlbums :: Request f (Tag -> APIKey -> Ready)

Get the top albums tagged by this tag, ordered by tag count.

Optional: limit, page

http://www.last.fm/api/show/tag.getTopAlbums

getTopArtists :: Request f (Tag -> APIKey -> Ready)

Get the top artists tagged by this tag, ordered by tag count.

Optional: limit, page

http://www.last.fm/api/show/tag.getTopArtists

getTopTags :: Request f (APIKey -> Ready)

Fetches the top global tags on Last.fm, sorted by popularity (number of times used)

http://www.last.fm/api/show/tag.getTopTags

getTopTracks :: Request f (Tag -> APIKey -> Ready)

Get the top tracks tagged by this tag, ordered by tag count.

Optional: limit, page

http://www.last.fm/api/show/tag.getTopTracks

getWeeklyArtistChart :: Request f (Tag -> APIKey -> Ready)

Get an artist chart for a tag, for a given date range. If no date range is supplied, it will return the most recent artist chart for this tag.

Optional: from, to, limit

http://www.last.fm/api/show/tag.getWeeklyArtistChart

getWeeklyChartList :: Request f (Tag -> APIKey -> Ready)

Get a list of available charts for this tag, expressed as date ranges which can be sent to the chart services.

http://www.last.fm/api/show/tag.getWeeklyChartList

search :: Request f (Tag -> APIKey -> Ready)

Search for a tag by name. Returns matches sorted by relevance.

Optional: limit, page

http://www.last.fm/api/show/tag.search