Network.Lastfm.Tasteometer

Description

Lastfm tasteometer API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Tasteometer as Tasteometer

Documentation

compare :: (Targeted u, Targeted v) => Request f Send u -> Request f Send v -> Request f Send (APIKey -> Ready)

Get a Tasteometer score from two inputs, along with a list of shared artists. If the input is a user some additional information is returned.

Optional: limit

http://www.lastfm.ru/api/show/tasteometer.compare