Network.Lastfm.User

Description

Lastfm user API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.User as User

Documentation

getArtistTracks :: Request f Send (User -> Artist -> APIKey -> Ready)

Get a list of tracks by a given artist scrobbled by this user , including scrobble time. Can be limited to specific timeranges, defaults to all time.

Optional: startTimestamp, page, endTimestamp

http://www.last.fm/api/show/user.getArtistTracks

getBannedTracks :: Request f Send (User -> APIKey -> Ready)

Returns the tracks banned by the user

Optional: limit, page

http://www.last.fm/api/show/user.getBannedTracks

getEvents :: Request f Send (User -> APIKey -> Ready)

Get a list of upcoming events that this user is attending. Easily integratable into calendars, using the ical standard (see 'more formats' section below).

Optional: page, festivalsonly, limit

http://www.last.fm/api/show/user.getEvents

getFriends :: Request f Send (User -> APIKey -> Ready)

Get a list of the user's friends on Last.fm.

Optional: recenttracks, limit, page

http://www.last.fm/api/show/user.getFriends

getInfo :: Request f Send (User -> APIKey -> Ready)

Get information about a user profile.

http://www.last.fm/api/show/user.getInfo

getLovedTracks :: Request f Send (User -> APIKey -> Ready)

Get the last 50 tracks loved by a user.

Optional: limit, page

http://www.last.fm/api/show/user.getLovedTracks

getNeighbours :: Request f Send (User -> APIKey -> Ready)

Get a list of a user's neighbours on Last.fm.

Optional: limit

http://www.last.fm/api/show/user.getNeighbours

getNewReleases :: Request f Send (User -> APIKey -> Ready)

Gets a list of forthcoming releases based on a user's musical taste.

Optional: userecs

http://www.last.fm/api/show/user.getNewReleases

getPastEvents :: Request f Send (User -> APIKey -> Ready)

Get a paginated list of all events a user has attended in the past.

Optional: page, limit

http://www.last.fm/api/show/user.getPastEvents

getPersonalTags :: Request f Send (User -> Tag -> TaggingType -> APIKey -> Ready)

Get the user's personal tags

Optional: taggingtype, limit, page

http://www.last.fm/api/show/user.getPersonalTags

getPlaylists :: Request f Send (User -> APIKey -> Ready)

Get a list of a user's playlists on Last.fm.

http://www.last.fm/api/show/user.getPlaylists

getRecentStations :: Request f Sign (User -> APIKey -> SessionKey -> Ready)

Get a list of the recent Stations listened to by this user.

Optional: limit, page

http://www.last.fm/api/show/user.getRecentStations

getRecentTracks :: Request f Send (User -> APIKey -> Ready)

Get a list of the recent tracks listened to by this user. Also includes the currently playing track with the nowplaying=true attribute if the user is currently listening.

Optional: limit, page, from, extended, to

http://www.last.fm/api/show/user.getRecentTracks

getRecommendedArtists :: Request f Sign (APIKey -> SessionKey -> Ready)

Get Last.fm artist recommendations for a user

Optional: page, limit

http://www.last.fm/api/show/user.getRecommendedArtists

getRecommendedEvents :: Request f Sign (APIKey -> SessionKey -> Ready)

Get a paginated list of all events recommended to a user by Last.fm, based on their listening profile.

Optional: limit, page, latitude, longitude, festivalsonly, country

http://www.last.fm/api/show/user.getRecommendedEvents

getShouts :: Request f Send (User -> APIKey -> Ready)

Get shouts for this user. Also available as an rss feed.

Optional: page, limit

http://www.last.fm/api/show/user.getShouts

getTopAlbums :: Request f Send (User -> APIKey -> Ready)

Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default.

Optional: period, limit, page

http://www.last.fm/api/show/user.getTopAlbums

getTopArtists :: Request f Send (User -> APIKey -> Ready)

Get the top artists listened to by a user. You can stipulate a time period. Sends the overall chart by default.

Optional: period, limit, page

http://www.last.fm/api/show/user.getTopArtists

getTopTags :: Request f Send (User -> APIKey -> Ready)

Get the top tags used by this user.

Optional: limit

http://www.last.fm/api/show/user.getTopTags

getTopTracks :: Request f Send (User -> APIKey -> Ready)

Get the top tracks listened to by a user. You can stipulate a time period. Sends the overall chart by default.

Optional: period, limit, page

http://www.last.fm/api/show/user.getTopTracks

getWeeklyAlbumChart :: Request f Send (User -> APIKey -> Ready)

Get an album chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent album chart for this user.

Optional: from, to

http://www.last.fm/api/show/user.getWeeklyAlbumChart

getWeeklyArtistChart :: Request f Send (User -> APIKey -> Ready)

Get an artist chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent artist chart for this user.

Optional: from, to

http://www.last.fm/api/show/user.getWeeklyArtistChart

getWeeklyChartList :: Request f Send (User -> APIKey -> Ready)

Get a list of available charts for this user, expressed as date ranges which can be sent to the chart services.

http://www.last.fm/api/show/user.getWeeklyChartList

getWeeklyTrackChart :: Request f Send (User -> APIKey -> Ready)

Get a track chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent track chart for this user.

Optional: from, to

http://www.last.fm/api/show/user.getWeeklyTrackChart

shout :: Request f Sign (User -> Message -> APIKey -> SessionKey -> Ready)

Shout on this user's shoutbox

http://www.last.fm/api/show/user.shout