Network.Lastfm.Venue

Description

Lastfm venue API

This module is intended to be imported qualified:

 import qualified Network.Lastfm.Venue as Venue

Documentation

getEvents :: Request f Send (Venue -> APIKey -> Ready)

Get a list of upcoming events at this venue.

Optional: festivalsonly

http://www.last.fm/api/show/venue.getEvents

getPastEvents :: Request f Send (Venue -> APIKey -> Ready)

Get a paginated list of all the events held at this venue in the past.

Optional: festivalsonly, page, limit

http://www.last.fm/api/show/venue.getPastEvents

search :: Request f Send (VenueName -> APIKey -> Ready)

Search for a venue by venue name

Optional: page, limit, country

http://www.last.fm/api/show/venue.search