Network.Lastfm.Internal

Description

liblastfm internals

You shouldn't need to import this module unless you are doing something interesting.

Documentation

newtype Request f a t

Lastfm API request data type

a is authentication state. Might be Send which indicates that you may send this request already or Sign, when request signature isn't computed yet

f is response format. liblastfm currently supports JSON or XML

Constructors

Request

Fields

unRequest :: Dual (Endo (R f a t))

data Format

Response format: either JSON or XML

Constructors

JSON

XML

data Auth

Authentication method

Constructors

Send
Public API. Doesn't require anything special besides API key

Sign
Private API. Requires Session key and Secret as well as API key

data Ready

Indicates that request is ready for sending

data R f a t

Lastfm API request data type

low-level representation

Constructors

R

Fields

_host :: !Text

_method :: !ByteString

_query :: !(Map Text Text)

wrap :: (R f a t -> R f a t) -> Request f a t

Wrapping to interesting Monoid (R -> R) instance

unwrap :: Request f a t -> R f a t -> R f a t

Unwrapping from interesting Monoid (R -> R) instance

class Coercing t where

Coerce requests changing their phantom parameters. Used to ensure right flow of working with liblastfm. If you use it on your worn, then you will break abstraction

Methods

coerce :: t (a :: Auth) b -> t c d

render :: R f a t -> String

Construct String from request for networking

Lenses

host :: Functor f => (Text -> f Text) -> R h a t -> f (R h a t)

Request _host lens

method :: Functor f => (ByteString -> f ByteString) -> R h a t -> f (R h a t)

Request http _method lens

query :: Functor f => (Map Text Text -> f (Map Text Text)) -> R h a t -> f (R h a t)

Request _query string lens