Network.Lastfm.Response

Description

Request sending and Response parsing

Sign Request

Signing is important part of every authentication requiring API request. Basically, every such request is appended with md5 footprint of its arguments as described at http://www.last.fm/api/authspec#8

newtype Secret

Application secret

Constructors

Secret Text

sign :: Secret -> Request f Sign -> Request f Ready

Sign Request with Secret

Get Response

class Supported f

Supported provides parsing for a chosen Format

JSON is parsed to aeson's Value, XML is to lazy ByteString (in other words, parsing XML is left to the user)

Associated Types

type Response f

lastfm :: Supported f => Request f Ready -> IO (Response f)

Send Request and parse the Response

lastfm_ :: Supported f => Request f Ready -> IO ()

Send Request without parsing the Response

Internal

lastfm' :: Supported f => (Proxy f -> ByteString -> ResponseHeaders -> a) -> R f -> IO a

Send R and parse Response with the supplied function

finalize :: Supported f => Request f Ready -> R f

Get R from Request